Added website documentation
Added the `docs/` folder, which contains the documentation in the shape
of a Quarto website.

The documentation is based on the pkgdown's version of the documentation
for the previous extension, acronymsdown. Elements that pertained to
RMarkdown exclusively were removed.

This documentation requires `knitr`, because the `styles` page uses bash
to render a preview for each style. Quarto is incapable of using bash
by itself, and must rely on either knitr, jupyter, or ObservableJS.
Knitr seems the best engine for this job (jupyter should be able to
use bash as well but is more focused on python); in addition, it
already has been proved to work, as it powered our previous
documentation (with pkgdown)!