Added workflow to build and publish documentation
This workflow is based on Quarto's example:
https://quarto.org/docs/publishing/github-pages.html

The workflow automatically builds the documentation (which lives in
the `docs/` folder), when a push is made to either the `master` or
`docs` branch. This allows us testing the workflow and the documentation
on this secondary `docs` branch before merging to `master` when the
results are satisfying. Ideally, pushing to `docs` should build the docs
in a separate folder, to avoid overwriting the "master-based" current
version...

We need to install `knitr` to build the `articles/styles` pages, and
thus R. We cannot use r-lib's action to install the R dependencies,
because we do not have a `DESCRIPTION` file (and it would be
overkill to have one...).
Instead, we simply use r-lib to install R, and use a brief R command
to install the dependencies. It might not be optimal (no cache,
not using binary packages...), but it should work.
We also install pandoc: it should be installed along Quarto, but
is not found when building our documentation, so just to be sure
we use r-lib's setup-pandoc action.