Refactored the 'styles' article to use the 'preview_style' script
The previous article used a Bash function directly in the document,
and executed this function for each style. It provided the correct
output... when no error happened. On the CI, if an error happens,
it does not provide any relevant information to debug and fix the
error.
Instead, this commit creates a 'preview_style.sh` script in the
'docs' folder, which serves the same purpose, but with more tests,
and error codes. An additional advantage is that it can be used
outside the 'styles' document, for example to preview a style when
localling developing new styles.
Also, added labels to the code chunks, so that we know which chunk
failed (it could be deduced from the argument given to the script,
but this was not fail-proof).