Fixed the website homepage
We want the (built) website homepage to be the same as the project's
README (at the root of the project). The previous version used a
*pre-render* script to copy this README to the docs folder, as
the `index.qmd` file. However, Quarto did not seem to acknowledge
this file correctly, and created by default a redirection to another
file.
To fix this, we create an almost empty `index.qmd` file, which simply
asks Quarto to include the content of the README when building the page.
This ensures that the homepage will always be up-to-date with the
README, and avoids tricking Quarto.