Refactored the Options into their own file
The table that holds the options was previously a local variable in the
"main file" (`parse-acronyms.lua`). To make it clearer, they are now
given their own "module" (the `acronyms_options.lua` file).

This commit is necessary to prepare for a more extensive refactoring,
moving some functions to their own files as well, and ensuring that
every file can import (`require`) exactly what they need, instead of
relying on global variables or functions. Some helper functions, in
particular, need access to the options.

It also prepares for more complex options, by using a table with
functions (i.e., a "class" in Lua world), instead of a simple table
with only data inside. For now, options are still accessed through
the old indexing way (`[]`); in the future, they could be obtained
through function calls. This would allow for context-specific
options, e.g., setting a style for a category of acronyms.