Re-added tests scenarii from acronymsdown
Re-added the tests, in the form of folders that each contain:
- an `input.qmd` file describing the Quarto document that should be
  rendered.
- an `expected.md` file describing the expected result, the rendered
  document. We use a markdown format, as this allows easy comparison
  (when compared to, e.g., HTML or PDF), even though Markdown is
  rarely used as the output format. Still, if the Markdown output
  is the same, we can assume an HTML version would be compliant as
  well.
- (optional) an `expected.stderr` file describing the expected error
  log. Note that error encompasses warnings as well (the name refers
  to stderr, the standard error file descriptor).
- (optional) a `Readme.md` file briefly describing the test.

These are the same tests as previously used in the acronymsdown package.

These are only "scenarii": alone, they do not provide a test suite.
They must be executed, and the output compared to the expectations,
in order to truly test that the extension is working.