Added a test script to run tests
Added the `run_tests.ts` script to automatically run each of the
tests "scenarii". The script calls Quarto with some options,
gets the output (stdout) and error (stderr), and compares them to
the scenario's expectations.

The test script prints a short result to the console (CLI) before /
after each test, to provide regular report to the user.
It may also, if the `GITHUB_STEP_SUMMARY` environment variable is set,
create a more detailed Markdown report, that can be visualized.
Locally, the report must be rendered to HTML, but on GitHub CI/CD it
can be automatically viewed when checking the result of a job.
This report contains, for each test: execution time, actual output /
expected output, actual error / expected error, input file.
In addition, it also contains the Quarto configuration, for full
reproducibility and debugging.

Tests can be selected through the command-line arguments.
For example, `quarto run tests/run_tests.ts 01-simple 02-custom-title`
will only run tests #01 and #02. By default (when no argument is
specified), all tests are performed.

Quarto options that are common to all tests (e.g., wanting the
ATX headers `# Header 1`) are specified in `_quarto.yml` to
simplify the test script. Additionnally, this ensures that these
options are applied even when calling Quarto without the test script!