Better formatting for the Md report
For some reason, GitHub does not like putting Markdown in a fenced code
block. The Markdown is not considered as "text" (to be written as-is),
but interpreted as Markdown.
However, using indented code blocks instead works: Markdown is not
interpreted. The only drawback is that there is no clear visual
indication of a code block (when compared to a fenced block),
and we cannot specify the language, so there is no syntax
highlighting.

Also, removed the heading slug in the test details (for example,
the `{#01-simple}` in `# 01-simple {#01-simple}`).
This is not parsed correctly by GitHub (appears on the header line),
so we remove it. The ability to link directly to a test might be
impaired...