Refactored acronyms styles to accept overriding parameter
The styles functions (`long-short`, `short-long`, and so on) determined
whether this was a first occurrence directly by calling the method
on the acronym object. This did not allow to override this behaviour for
a specific instance (a feature that can be added thanks to Quarto
shortcodes' keyworded-arguments).

This commit thus adds a new parameter `is_first_use` for most of these
styles functions (except those that do not differentiate between first
and next uses), as well as for the "public API" `replaceWithStyle`
function. For this function, the parameter can be set to `nil`, in which
case the correct value will be determined by a call to
`acronym:isFirstUse`, just like before.

The commit thus keeps consistency with previous versions (no breaking
changes), but prepares for a new feature.