Simplified require
We previously tried to find the path to the script's directory
in order to add it to the `package_path`, so that the `require()`
calls can find the other files in the same directory (which is not
the current working dir!).

However, this was very messy and not resilient at all (using the
`debug` module, and parsing the path...).
According to Issue #8 this did not work on Windows for some reason
(the path found for the current script was `nil`).

This commit tries to simply require the files, without changing
the `package_path`. As they all are in the same folder, this
should work.