Added option to customize LoA header classes
The new option `loa_header_classes` should be a list of additional
classes that the Header (title) will have.

Even though this behaviour could already be obtained by disabling the
Header (`loa_title` set to `""`) and by manually putting a Header in
the qmd document, this option will simplify customizing the Header.

This change was prompted by Issue #6.