Added test for LoA extra classes
This new test focuses on the new `loa_header_classes` option, and
particularly tries to put the `.unnumbered` class on the Header, to
make it an unnumbered section.

The behaviour when no extra classes are defined should be tested by
all other tests (because this option uses an empty table by default).

Setting multiple extra classes should also work, but because this would
not be very useful in our context (unless we define the meaning of these
extra classes), it is not part of the automated test suite.