Fixed imports in the test script
We imported functions from "sub-modules" files, e.g.,
`streams/conversion.ts`; this was not resilient to Deno changes,
as the files can be renamed. It is better to import from the module
itself, in this case `streams/mod.ts`.

This prepares the transition to Quarto's next release, which uses
a more recent (although not the latest) version of the Deno stdlib.
In particular, the `writeAllSync` was in `streams/conversion.ts` in
Deno@166.0, but in `streams/write_all.ts` in Deno@185.0.
Failing to find this function made our test suite crash with
Quarto 1.4.

An alternative would have been to dynamically load the function from
one of these paths, using `try/catch` and the `import()` function.
However, it is much more verbose, and less easy to read (declaring
a mutable variable, trying to load the module and assigning the
function, trying another path if an error is caught, ...).