Write test result to file instead of stdout
We previously wanted the resulting document of each test to be
printed to stdout (the standard output), so we can access it easily.
This worked well on Linux and macOS, but on Windows it produces an
error when using Quarto before v1.4 (incorrect temporary path).

To make sure the testing workflow works on Windows, we now write
the resulting document to a file.
This requires changing the way we filter the stderr (standard error)
as well, because Quarto also emits the filename in this case.