Refactored test to include additional details in the summary report
Because Quarto produces additional lines to the stdout and stderr,
we needed to filter them to compare to the expected results.
However, we only reported the filtered versions to the (GitHub)
summary report; if our filter does not work, the test seems to
fail and we do not understand why.

Thus, this commit removes the `outputLines` (and similar variables),
which were not necessary (we can derive them anyway from other
variables), and adds the `originalOutput` and `originalError` variables.