Fix failed tests on Windows
Tests were failing specifically on Windows because we parse the output
file with `\n` (the LF line separator); however, Windows uses `\r\n`.
Thus, the result array contained `\r` at the end of each line, and
the comparison did not match (`---` != `---\r`).
We now use a regexp that checks for the return carriage, and should
account for both CRLF and LF separators.