Fixed the preview_style when generating docs
A previous commit changed the way we `require` Lua files, in order
to fix a bug on Windows. However, the new way does not set the path
automatically. When rendering through Quarto, this is not a problem,
as Quarto helps with loading files in an extension folder.

Yet, the `preview_style` script does not use Quarto, because it is
complicated to get the extensions folder and to render a temporary
file. Pandoc is far easier for this script, but it does not set the
path, which provoked an error.

This commit fixes the preview script so that we `cd` to the extensions
folder just before we render the document; all Lua files can be loaded
correctly by doing so.