Fixed the testing workflow
We previously used the value of `''` (empty string) to download the
latest release of Quarto, so that we could then use `pre-release`
for the pre-release.

However, Quarto changed their setup workflow, the empty string is
not taken as-is, which means we tried to download the version `v`
(which does not exist).
We must now use `release` for the release (latest stable) version.