Added an example for multi-document projects
As per issue #11, the default configuration of acronyms does not work
very well on multi-document projects, especially websites and books.
This is vastly due to the way Quarto renders each file separately,
but can be mitigated by adjusting acronyms' configuration.

This new documentation page demonstrates two different setups for such
multi-document projects: one with a single List of Acronyms, and one
with a separate list on each page.

In addition, refactored the previous `advanced_usage.qmd` article into
a new folder ("Advanced usage" in the menu), and splitted the sections
into separate articles in this folder. This is because the "Advanced"
articles can become quite long, and would not be very readable on the
same page.