Better error message for misconfigured acronyms
When an acronym's shortname and/or longname was not specified correctly,
the acronyms filter would raise an error, but this error was not very
descriptive of the problem. It looked more like an internal error,
rather than a user error.

Added a new function that generates a more verbose error message, based
on the Tidyverse recommendations (because they make for quite clear and
actionable errors, indicating what is the problem and how we can solve
it).
See https://style.tidyverse.org/error-messages.html

Also added an helper function that is used to show how the acronym was
defined in the metadata (by simply dumping the table). This is useful
to identify which acronym was incorrectly created (previous error
message would just say "shortname cannot be nil!", which does not inform
us of which acronym is faulty...).

This change was particularly prompted by Issue #12, which highlighted
that error messages were not very clear for users.