Fixed docs and added listings for advanced usages
Several of the docs links were out-of-date since we added the "advanced"
folder with tutorials inside, instead of a single "advanced" page in the
"articles" folder.

This commit fixes these links, and adds a new page inside the "advanced"
folder, which is a listing of all existing tutorials inside this folder.
This does not replace the dropdown menu in the navigation bar, but
allows linking to a separate "advanced usage" page, as if we were
linking to the folder as a whole.