Refactored "current usage order" to the Acronyms table
Previously, the `parse-acronyms.lua` script was responsible for setting
the acronyms' usage order, when an acronym appears.
This feature is now moved to the `Acronyms` table, so that the necessary
variables are grouped together (rather than relying on script-local
variables), although it is still invoked from `parse-acronyms.lua`.
This change will allow to re-use the same functionality when using
shortcodes, without repeating the code between several scripts.