Added keyword arguments for shortcodes
This allows shortcodes to override the parameters defined in the
`Options` table, which is parsed from the metadata.

For example, a specific acronym can be replaced with a style that
differs from the others acronyms, by specifying the `style=...` keyword
argument when using the shortcode.

Strangely, the keyworded arguments return an empty Inlines when they are
not specified, rather than `nil`. When they are specified, they return
a non-empty Inlines. This means that we must take care when handling
them, first to check whether they were specified or not, and secondly to
return their string representation if they exist.