Allow tests to have non-existing files
In the automatic test suite, we read several files to compare the output
of Quarto to the expected result, the potential error messages, etc.
However, when one of these files does not exist, mainly in case of
errors during the render, the test suite crashes because we do not
expect such files not to exist.

In such crashes, it can be difficult to debug the cause, because the
test suite completely stops.
This commit avoids such crashes by using a default empty string if these
files cannot be read. It does not change the test outcome, because the
comparison of an expected (non-empty) file and an empty string would
fail anyway, but the test suite keeps working.

Additionally, it will allow to create tests based on (voluntarily)
raising errors. There would be no output document, thus a default empty
string, compared to an empty expected document (because in these cases
we know that there will be no output), which works.