Fixed keyworded arguments in shortcodes
Keyworded arguments are obtained as Pandoc Inlines objects, and we
parsed them to strings. However, in some cases (for example,
`first_use`, `insert_links`, ...), we want to get a boolean value.
When using a string, Lua often assumes a truthy value (if the string
is not empty, etc.), which is not what we want!

This commit adds a new helper function that converts common strings
(`"true"`, `"false"`, `"yes"`, `"no"`, `"y"`, `"n"`) into their
corresponding boolean value.
The shortcode functions are also updated to get their keyworded
arguments as boolean (by using this new helper) when necessary.

Also fixed the way we handle default values for such arguments:
we previously used `value = value or Options["value"]`, which works
fine with nil/non-nil values, but wrongly considers a `false` value
as an empty value (and replaces it with the default)!
We keep the `x = x or ...` for strings, but replace with an if for
boolean values.