Fixed docs: load acronyms extension to preview styles
Previewing the styles was always kind of a hack between Quarto and
Pandoc. Since we directly use Quarto libraries (`log`) in our extension,
the preview was even harder.

This commit simplifies the workflow by leveraging the shortcodes: we now
load the extension by creating a symlink automatically when building the
docs (similar to what we do in tests), which allows us to use shortcodes
in articles.

Instead of using an external preview script that calls another instance
of Pandoc, the `styles` article now directly renders the desired style
with shortcodes.

However, shortcodes are now enabled on all pages of the docs; in a very
irritating manner, Quarto still parses and renders them even when they
are in code-blocks or inline code. For example, writing: "Type `{{< acr
KEY >}}` to use acronyms" will render as "Type KEY to use acronyms",
which is not what we want.

For code-blocks, we can use the `{shortcodes=false}` option, but for
inline codes we have to encapsulate them in triple braces `{{{`.