Add the Translations module
This module is made to support some (given) sentences, according to the
user's preferred language. It contains both the sentences to translate,
and the algorithm to identify the best matching translation for a given
language, following the relevant standards (RFC4647 and RFC5646).

Note: it may not be optimal in terms of performance, as we first build
the table of all languages to match (for example, `zh-Hant-CN` can be
matched to `zh-Hant-CN` itself, `zh-Hant`, or `zh` (in this order).
If the requested language contains many such subtags, it will take more
time; and if, in addition, we have an exact match, this table will have
been built for "nothing"! However, that scenario should not appear very
frequently in Quarto. Most people will use a "simple" language (e.g.,
`en` or `fr`), and thus the table-construction step will be short.