Use the translated LoA title when generating the LoA
We previously relied only on the `loa_title` option (in the Options
module), and the shortcode parameter (when called from a shortcode).
The `loa_title` option was set to a sane default, in case it was not
specified by the user.

However, we now want to use the Translations module when the user does
not specify a title. This is done by setting the `loa_title` option to
`nil` by default, so we can recognize when an option is not set (instead
of the previous default).
Remember that we must use Pandoc elements; the metadata options and
shortcode parameters are both `Inlines` containing a `Str`, we keep them
as-is when they are set; but we must enclose our translation into the
same elements otherwise.