Add support for custom list of acronyms format
We previously rendered the List of Acronyms (LoA) as a DefinitionList.
This is a somewhat standard element for such lists: names and definitions.
However, people may want to customize the format, see for example issue #19.
When rendering to HTML, the DefinitionList can be styled thanks to CSS,
but that does not work on other outputs (PDF, DOCX, ...).

This commit:
* Adds a `loa_format` option to specify how the LoA should be rendered.
  Because the YAML metadata are parsed by Pandoc, we must use the
  special "raw" syntax (`{=raw}`) when we want to to specify a Markdown
  template (for example, `**{shortname}**` to have the shortname in
  bold). It also means we should parse it in a special way.
* Refactors the way we generate the LoA to support both the
  DefinitionList (default behaviour) and the custom Markdown rendering.