Fix the custom LoA format for multiple blocks
The implemented feature previously assumed a single "block" (in Pandoc
terms) rendered for the List Of Acronyms when using a custom template.
This works perfectly with, e.g., a bullet list (it is a single "block"
consisting of multiple inline elements), but not when each acronym is
rendered as its own block (such as a paragraph).

This commit fixes this, by wrapping the result in a Div (a single
block). However, this changes the result for any given template;
the previous test thus had to be updated.