Add support for non-alphanum characters in shortnames
Because shortnames are used as default keys, and keys are used as
anchors when linking to the List Of Acronyms, using any
non-alpha-numerical character would make the anchor fail, as highlighted
by #31

This commit adds sanitization of keys when creating the anchor, by
replacing any non-ASCII character (and, even stricter, any character
that is not a digit (0-9), a lowercase non-accented letter (a-z), or an
uppercase non-accented character (A-Z)) with underscores '_'.

Note that it is not recommended to use accents in the shortname (or even
worse, only accents), because of the way Lua handles strings: they are
replaced with 2 underscores '__'.