Add support for "simplified" format (`fromfile` documents)
External documents (`fromfile`) may define acronyms outside the main
Quarto document. We initially used the same code to parse such acronyms
as in the main document, which meant they needed to use the
`acronyms.keys` structure, which is quite verbose.

A new, simplified format was proposed in #33 ; this commit implements
the parsing of such files, while retaining compatibility with previous
versions.