Add support for plural form when parsing acronyms
While some acronyms have a trivial plural form (e.g., adding a single s
at the end of the word), there are many that must be defined by the
user, because they use non-trivial rules. Examples include fish,
stimuli, ...

This commit parses an optional `plural` block from the YAML, and uses
the default rule of adding a final s instead if this block is not
defined, or if one of short name and long name is not defined.