Add support for plural syntax when writing a document
The function that parses "commands" (`replaceAcronym`) is changed to
support both `\acr` and `\acrs`, the latter requesting the plural form.
The functions that process the acronym rendering are changed to support
this new argument, and ultimately the `acronyms_styles` module uses the
plural data from the acronym definition when rendering if necessary.