Fix handling of symbolic link when building docs
When building the docs, we created a symbolic link to the _extensions
folder, so that the documentation can use the acronyms to demonstrate
its use.

However, there was a bug that created a symbolic link inside the
_extensions directory, pointing to itself in an unfinite loop; this bug
is solved by adding the `-n` option to `ln`, to avoid following the link
when creating the new link. In addition with the `-f` option, this means
that we simply do not create a link if it already exists.

Note that, on macOS, this is usually the `-h` option; but that option
does not exist on other versions of `ln` (e.g., the one available on
Ubuntu). To improve chances of this command working on other systems,
`-n` is used...

The `ln` manual states that both `-h` and `-n` are non-standard...