Update filter to also support simplified commands
Same as for the shortcodes, the filter now supports:
- ACR (upper case);
- ACRS (upper case + plural).

Note that, because of how I implemented the regex, an user could use
a bizarre command such as `AcR` (or any variation). Technically, this
would count in the `else` part, as if it was `acr`.

I'm not sure whether it would hinder the performance to ensure that only
`acr`, `Acr`, or `ACR` can be used. Because the user should not try this
anyway, I don't think it would be worth it.

                -- Technically, this `else` is also triggered when the command
                -- is `AcR` (or any variation like that). I'm not sure whether
                -- it would be better, in terms of performance, to ensure that
                -- only `acr`, `Acr`, or `ACR` can be used.